feat(quality): スコア詳細ダイアログを新UIに統合し、旧UIを削除
PR-2（#293）の as-is 取り込みで、スコア詳細ダイアログが2種類並立していた。

  /quality                   新UI（app/quality/page.tsx 内にインライン687行・政策評価あり）
  /sankey-svg                旧UI（共有 ScoreDetailDialog 515行・政策評価なし）
  /subcontracts/[projectId]  旧UI（同上）

新UIを共有コンポーネントへ移して旧UIを置き換え、3ページとも同じ表示にする。

## /api/policy-summary に pid モードを追加

新UIは PolicyEvaluation を26フィールド使うが、サマリの圧縮形（o/d/e/t/x/n/r/a/c）
では recommendationReason・findings などを返せない。?pid= 指定で完全な
PolicyEvaluation を返すモードを足した。

政策評価は母集団のパーセンタイル・分位点から閾値を決めるため1事業だけでは
算出できない。全件評価を年度ごとに1回組み立ててキャッシュし、サマリと pid 単体で
使い回す（buildEvaluations）。存在しない pid は 404。

## コンポーネント構成

  client/components/quality/score-meta.tsx        新設241行
      軸の文言・重み・説明、推奨判断/改善アクションのバッジ、支出先ステータス。
      /quality の一覧とダイアログで共用（従来はページ内に重複していた）
  client/components/quality/ScoreDetailDialog.tsx 旧515行 → 新772行
      支出先・事業詳細に加えて政策評価も自前で取得する。policy を props で
      受け取れば取りに行かない（/quality は全件から算出済みのものを渡す）

## 削除

- client/hooks/useScoreDetailData.ts（ダイアログが自前取得するため不要）
- app/quality/page.tsx の formatAmount / pct / scoreColor
  （既存の共有 score-format.ts と同一実装の重複だった）

app/quality/page.tsx は 1,994行 → 1,024行。

検証: tsc クリーン / lint エラーなし / vitest 100件パス / build 成功 /
dev で5ページ 200・ダイアログが引く3経路（policy-summary?pid / recipients /
project-details）すべて 200
